refactor(endless): move endless memory services to extension
