+++ src/extensions/endless-memory/index.ts
@@ -1,0 +1,1 @@
+export * from './endless-memory-services.js';
